Feat(multi-window): Introduce window navigation keys and better window kayout
This commit introduces a big level of complexity
over the window management system. Now the layout
of the windows is stored in a binary tree, so that
the computation of the visual window layout is
clener and more correct.

Keybindings:
CTRL_W now opens the window mode, which triggers an
action based on the consequently pressed key:
- ARROW_LEFT / h = switch to the window to the left
- ARROW_RIGHT / l = switch to the window to the right
- ARROW_UP / k = switch to the window on top
- ARROW_DOWN / j = switch to the window on bottom
- q = close the active window
- | = open a new window splitting the screen vertically
- - = open a new window splitting the screen horizontally